requirements.txt
feedparser==6.0.11
urllib3==2.2.3
psycopg[binary]==3.2.1
SQLAlchemy==2.0.36
python-dotenv==1.0.1
pandas==2.2.3
scikit-learn==1.5.2
sentence-transformers==3.0.1
tqdm==4.66.5
numpy==2.1.1
